requirements.txt
pytest~=7.4.2
selenium~=4.14.0
allure-pytest~=2.13.2